new rule: enforce `name` getter on custom elements
Most minifiers will rename variables, classes and functions to a single
letter, or more if the alphabet has been exhausted. This means that any
call to `Function.name` will result in unwanted results.

This patch adds a rule that applies to custom elements. The rule
enforces the existence of a `static get name()` method that returns the
name of the class as a string. Ensuring that any call to `.name` on the
custom element will result in the correct name.

Co-authored-by: Keith Cirkel <keithamus@users.noreply.github.com>