generalize `filenames/match-regex` to allow any 'extension prefix'
For example, `file-name.server.tsx`, `file-name.d.ts`, `file-name.config.json`

This allows any of the following:

file-name
file-name.d
file-name.config
file-name.server
file-name.wild-card

Note that the file extension is not part of the regex.

I made this change because we are seeing more and more of this style of 'extension prefix` convention.